refactor: step3 외부 참조 없는 변수 캡슐화, 테스트용 파일 위치 변경
