+++ README.md
@@ -10,6 +10,15 @@
 
 
 # STEP3 TODO
+- [ ] 클래스변수와 멤버 변수 개행 분리
+- [ ] 자동차 스스로 움직일지 말지 결정 
+- [ ] carController 이름과 역할 바꿔보기 
+- [ ] inputView 책임에 맞게 print 문 변경하기
+- [ ] 게임을 실행하는 역할에 맞게 Round 이름 변경해보기
+- [ ] Rount init을 생성자에서 처리하기
+- [ ] step3 패키지 이름 변경하기
+
+#  STEP3 DONE
 - [x] 자동차 이동
   - [x] 주어진 횟수 동안 n대의 자동차는 전진 또는 멈출 수 있다.
   - [x] 전진하는 조건은 0에서 9 사이에서 random 값을 구한 후 random 값이 4이상일 경우이다.
@@ -17,11 +26,12 @@
   - [x] 사용자는 몇 대의 자동차로 몇 번의 이동을 할 것인지를 입력할 수 있어야 한다.
   - [x] 자동차의 상태를 화면에 출력한다. 어느 시점에 출력할 것인지에 대한 제약은 없다.
   - [x] UI 로직을 InputView, ResultView와 같은 클래스를 추가해 분리한다.
-- [x] 테스트 
+- [x] 테스트
   - [x] 모든 로직에 단위 테스트를 구현한다
-- [x] 구조 & 컨벤션 
+- [x] 구조 & 컨벤션
   - [x] 핵심 로직을 구현하는 코드와 UI를 담당하는 로직을 구분한다.
   - [x] 자바 코드 컨벤션을 지키면서 프로그래밍한다
   - [x] else 예약어를 쓰지 않는다
-
-#  STEP3 DONE
+- [ ] 피드백
+  - [x] 외부 참조가 없는 변수 private
+  - [x] 테스트을 위한 파일은 테스트 패키지에만 두기
