+++ src/main/java/step3/car/Car.java
@@ -1,7 +1,8 @@
 package step3.car;
 
 public class Car {
-    public static final int DEFAULT_POSITION = 0;
+    private static final int DEFAULT_POSITION = 0;
+    
     private int position;
 
     public Car() {
